Highly Complete: Clean up warnings and error check
More error checking, and better initialization, as well as significantly
reducing stack usage during the player initialization.

+++ Frameworks/vio2sf/vio2sf/include/vio2sf/ARMJIT_A64/ARMJIT_Compiler_A64.h
@@ -231,7 +231,7 @@ public:
 
     u32 SubEntryOffset(JitBlockEntry entry)
     {
-        return (u8*)entry - GetRXBase();
+        return (u32)((u8*)entry - GetRXBase());
     }
 
     bool IsJITFault(const u8* pc);
